# [MM-30141] Break out of loop early in ClearSessionCacheForUserSkipClusterSend