@streamer45 2020-10-30
Interesting, tests are failing. Maybe the full loop is needed after all. Will check.