@mgdelacroix 2020-10-30
Just tested the server code and we can have multiple sessions cached, so I think it is indeed needed :/