@streamer45 2020-10-30
> Just tested the server code and we can have multiple sessions cached, so I think it is indeed needed :/

Thanks @mgdelacroix . This approach will not work then. Maybe we could do something at the cache level to mitigate a bit the locking in this case but it will need more thinking. I'll close this.