@agnivade 2020-10-30
Ah right, it's cached by the session id, not the user id. Good catch @mgdelacroix !